Split road_render.rs (869 lines) into road_render/ subdirectory
Break the monolithic road rendering module into focused submodules:
- mod.rs (232 lines): components, re-exports, main sync system
- tessellation.rs (186 lines): road segment mesh tessellation
- lane_markings.rs (376 lines): dashed/solid lane marking generation
- intersection.rs (80 lines): intersection disc mesh builder
- bezier.rs (17 lines): Bezier curve evaluation helpers

All pub items re-exported from mod.rs; no API changes for consumers.

Part of #1389
